Move ffi config to drop_types
