+++ drop-types/Cargo.toml
@@ -4,3 +4,5 @@ version = "1.0.0"
 edition = "2021"
 
 [dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
